requirements.txt
PySocks==1.5.6
requests==2.9.1
termcolor==1.1.0
Flask==0.10.1

